Apply comments
