+++ src/core/tests/eval.cpp
@@ -537,10 +537,38 @@ TEST(eval, evaluate_broadcast_v3_explicit_dyn) {
     ASSERT_EQ(result_val, expec);
 }
 
+class TestOpMultiOut : public op::Op {
+public:
+    OPENVINO_OP("TestOpMultiOut");
+    TestOpMultiOut() = default;
+
+    TestOpMultiOut(const Output<Node>& output_1, const Output<Node>& output_2) : Op({output_1, output_2}) {
+        validate_and_infer_types();
+    }
+
+    void validate_and_infer_types() override {
+        set_output_size(2);
+        set_output_type(0, get_input_element_type(0), get_input_partial_shape(0));
+        set_output_type(1, get_input_element_type(1), get_input_partial_shape(1));
+    }
+
+    std::shared_ptr<Node> clone_with_new_inputs(const OutputVector& new_args) const override {
+        return std::make_shared<TestOpMultiOut>(new_args.at(0), new_args.at(1));
+    }
+
+    OPENVINO_SUPPRESS_DEPRECATED_START
+    bool evaluate(const HostTensorVector& outputs, const HostTensorVector& inputs) const override {
+        inputs[0]->read(outputs[0]->get_data_ptr(), inputs[0]->get_size_in_bytes());
+        inputs[1]->read(outputs[1]->get_data_ptr(), inputs[1]->get_size_in_bytes());
+        return true;
+    }
+    OPENVINO_SUPPRESS_DEPRECATED_END
+};
+
 TEST(eval, test_op_multi_out) {
     auto p = make_shared<op::Parameter>(element::f32, PartialShape{2, 3});
     auto p2 = make_shared<op::Parameter>(element::f64, PartialShape{2, 2});
-    auto so = make_shared<ov::TestOpMultiOut>(p, p2);
+    auto so = make_shared<TestOpMultiOut>(p, p2);
     auto fun = make_shared<Function>(OutputVector{so->output(0), so->output(1)}, ParameterVector{p, p2});
     auto result = make_shared<HostTensor>(element::Type_t::f32, Shape{2, 3});
     auto result2 = make_shared<HostTensor>(element::Type_t::f64, Shape{2, 2});
